feat: authentication

+++ actions/banner.ts
@@ -7,21 +7,35 @@ import { cookies } from 'next/headers'
 export async function createBanner(formData: FormData) {
    const language = cookies().get('language')?.value as string
    const values = {
-      title: formData.get('title') as string,
-      imageUrl: formData.get('imageUrl') as string
+      imageUrl: formData.get('file') as string
    }
-   if (!values.imageUrl || !values.title) {
+   if (!values.imageUrl) {
       return {
          status: Status.BadRequest,
          message: 'Chưa nhập đủ thông tin'
       }
    }
+   const imageForm = new FormData()
+   imageForm.append('file', formData.get('file') as Blob)
+   imageForm.append('typePath', 'iit')
 
    try {
+      const result = await fetch(process.env.NEXT_APP_API_URL as string, {
+         method: 'POST',
+         body: imageForm
+      })
+      if (!result.ok) {
+         return {
+            status: Status.InternalServerError,
+            message: 'Co loi xay ra, khi upload anh!'
+         }
+      }
+      const data = await result.json()
+
       await prisma.banner.create({
          data: {
-            imageUrl: values?.imageUrl,
-            name: values.title
+            imageUrl: data?.fileName,
+            name: 'banner'
          }
       })
       revalidatePath(`/${language}/admin/banner`)
@@ -49,7 +63,19 @@ export async function getBanners() {
 
 export async function deleteBanner(id: string) {
    const language = cookies().get('language')?.value as string
+
    try {
+      // Đếm số lượng banner hiện có
+      const bannerCount = await prisma.banner.count()
+
+      // Nếu chỉ còn 1 banner, ngăn việc xóa
+      if (bannerCount <= 1) {
+         return {
+            status: Status.BadRequest,
+            message: 'Không thể xóa, chỉ còn một banner duy nhất!'
+         }
+      }
+
       await prisma.banner.delete({
          where: {
             id
